Create willYouMakeIt.java
This program will take in your cars MPG, gas in the tank, and the distance to the gas station and tell you if you will make it or not. Then, it will tell you how much you would need to make the trip if you wouldn't already or tell you how much further you could go without gassing up.